api-gateway/requirements.txt
Flask~=1.1.2
Jinja2~=2.11.3
MarkupSafe==1.1.1
itsdangerous==1.1.0
Werkzeug==1.0.1
requests
flask_cors
pyjwt==1.4.2